requirements.txt
torch>1.8.0
tensorboard
datasets==1.5.0
tensorflow==2.4.0
transformers==4.4.0
tokenizers<=0.10.2
hydra-core==1.0.0
matplotlib
tqdm
soynlp
seaborn
Flask
Flask_restful
tensorflow_addons